third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-angle-expected.txt

This is a testharness.js-based test.
[FAIL] Animating a custom property of type <angle> with iterationComposite
  assert_equals: expected "250deg" but got "50deg"
Harness: the test ran to completion.
